# feat: handle various month and year durations