--- a/beancount_interpolate/common.py
+++ b/beancount_interpolate/common.py
@@ -1,5 +1,6 @@
 import datetime
 from dateutil.relativedelta import relativedelta
+import math
 import re
 from beancount.core.number import D
 from beancount.core.amount import Amount, mul
@@ -26,21 +27,27 @@ def extract_mark_posting(posting, config):
     return False
 
 
-def get_number_of_txn(begin_date, duration, step, max_txn=float('inf')):
+def get_number_of_txn(begin_date, duration, step):
     """
     Computes the number of transactions within a given interval and step length.
 
-    Note: As the current implementation is a bit costly with a very large duration
-    and a small step size there is a `max_txn` parameter which can be set.
-    This is meant to be used with the config `max_new_txn`.
-    Upon exceeding `max_txn` the function returns `max_txn + 1`.
-    """
-    n_txn = 0
+    Note: This implementation requires 'step' to contain unique units which means either one of 'years', 'months' or
+    'days'. Something like e.g. 'months=2, days=3' is not supported and might lead to wrong results.
+    """    
+    
     end_date = begin_date + duration
-    while begin_date + n_txn * step < end_date:
-        n_txn += 1
-        if n_txn > max_txn:
-            break
+    diff = relativedelta(end_date, begin_date)
+    if step.years:
+        n_txn = math.floor(diff.years / step.years)
+    elif step.months:
+        diff_months = diff.months + 12 * diff.years
+        n_txn = math.floor(diff_months / step.months)
+    elif step.days:
+        diff_days = (end_date - begin_date).days
+        n_txn = math.floor(diff_days / step.days)
+    else:
+        raise ValueError(f"Unsupported step size of '{step}'. Units of 'years', 'months', 'days' have to be explicit")
+
     return n_txn
 
 
@@ -126,7 +133,7 @@ def distribute_over_period(params, default_date, total_value, config):
     """
 
     begin_date, duration, step = parse_mark(params, default_date, config)
-    period = get_number_of_txn(begin_date, duration, step, max_txn=config['max_new_tx'])
+    period = get_number_of_txn(begin_date, duration, step)
 
     if(period > config['max_new_tx']):
         period = config['max_new_tx']