This PR fixes some problems with durations and step sizes being specified as number of days. This leads to problems in leap year or months with different number of days e.g.

This implementation improves on this by utilizing `dateutil.relativedelta.relativedelta`, which handles such corner cases nicely. Additionally the maximum value for `max`, `inf` and `infinite` for the duration value is getting set by using `datetime.MAXYEAR`. This was necessary as adding `relativedelta` objects can lead to `OverflowError`.