@Akuukis 2022-03-20
Thanks!

Apologies for master branch - it turns out I've forgot to push last changes. You should be able to rebase without conflicts.

I tried out the PR on my files and got plenty of errors like these (shortlist from 500+ lines):
```
WARNING: Using defaults, because cannot parse mark "3": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "3": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "5@2021-04-01": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "7": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "30": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "720": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "60 @ 2021-11-01": unsupported operand type(s) for +: 'datetime.date' and 'int'
WARNING: Using defaults, because cannot parse mark "1@2022-03-03": unsupported operand type(s) for +: 'datetime.date' and 'int'
```

Looks like numeric length is not handled anymore. See commit `ab05d0b`, I added a regression test for this.