@RonMcKay 2022-03-20
Ah, did not know that just a numeric specification should also be supported. There was nothing in the README mentioning this. I will rebase and fix that.