@RonMcKay 2022-03-20
> Overall it works good 👍 See comments inline.
> 
> Little concern is that I had few new balance errors in the ledger, but they were only on few already otherwise questionable accounts and accounting practices, so that's fine.
> 
> Indeed, README doesn't mention plain numbers. Do you agree with a paragraph like this?
> 
> ```gfm
> The content of meta tag consists of 4 parts, every of them optional:
> 1. amount, a number. Defaults to 1.
> 2. period, a case-insensitive keyword "day", "week", "month" or "year". Defaults to "day".
> 3. start date, prefixed with `@` sign. Defaults to transaction date.
> 4. step size, prefixed with `/` sign. Can be either keyword from as period. Defaults to "day".
> 
> Summary: `[x] [period] [@ YYYY-MM[-DD]] [/ step]`
> ```

This is currently not correct as one can not combine `[x]` and `[period]`. It rather has to be `[x|period]` as in the code above `RE_PARSING`. When I open the second PR where I implemented the more flexible duration / step then we can adapt the README like that.

I pushed another commit where I improved the performance of the `get_number_of_txn` function by utilizing that we only have unique units as a step size. Can you try that out on your ledger? Should be way faster than before. This change also answers the comments regarding `if(period > config['max_new_tx'])` as the function itself does no longer handle this and the `if` condition can stay.