@Akuukis 2022-03-21
Right, I fixed the readme and pushed to master. As part of the next PR, please adjust it back.

Code looks good. I timed it again 3 times and it's faster but anyways slow at 21.4 to 21.9 seconds, that's still an ~40% increase.

I poked around with some experiments, and found put that this change on top of your PR makes everything fast again at 16.5 to 17.3 seconds.

![image](https://user-images.githubusercontent.com/7459419/159219538-79d77443-a415-41af-bf3b-52936c36203a.png)