@RonMcKay 2022-03-21
ah I see, thanks for investigating. The recomputing of the `end_date` and the step in the `while` condition is indeed a waste of time by me.
The reason why I recomputed the `date` everytime starting from `begin_date` was the following:

Here we can see that starting at 31.01.2022 and going two months forward by calling `begin_date + 2 * rd` results in the 31.03.2022.
```
>>> from dateutil.relativedelta import relativedelta
>>> import datetime
>>> begin_date = datetime.datetime(2022, 1, 31).date()
>>> begin_date
datetime.date(2022, 1, 31)
>>> rd = relativedelta(months=+1)
>>> begin_date + 2 * rd
datetime.date(2022, 3, 31)
```

But if we instead overwrite a date, we get the 28.03.2022.
```
>>> date = begin_date + rd
>>> date
datetime.date(2022, 2, 28)
>>> date = date + rd
>>> date
datetime.date(2022, 3, 28)
```

I currently do not have an idea if this can cause any problems but if this speeds up the code significantly then we could leave it like that.

Could you test the speed of the following code snippet:

```
i = 0
end_date = begin_date + duration
tmp_date = begin_date + i * step
while tmp_date < end_date and tmp_date <= datetime.date.today():
    accumulated_remainder += total_value / period
    if(abs(round_to(accumulated_remainder)) >= abs(round_to(config['min_value']))):
        amount = D(str(round_to(accumulated_remainder)))
        accumulated_remainder -= amount
        amounts.append(amount)
        dates.append(tmp_date)
    i += 1
    tmp_date = begin_date + i * step
    if(tmp_date > datetime.date.today()):
        break
```

If the speed of this is acceptable then it could be a compromise between the two.