@Akuukis 2022-03-21
Tested, this is as fast as previously :+1: 

Unrelated note, but this seems as a duplicate for `and tmp_date <= datetime.date.today():`, please remove it along the way:
```python
    if(tmp_date > datetime.date.today()):
        break
```