@RonMcKay 2022-03-21
I added the commit.

Additionally I replaced `tmp_date <= datetime.date.today()` by `tmp_date <= today_date` and defined `today_date` before the loop. This might save a few ms extra.